flask_app/requirements.txt
Flask==2.2.2
requests==2.28.1
gspread==5.4.0
oauth2client==4.1.3
haystack==1.5.0
transformers==4.30.0
tokenizers==0.13.3
cachetools==5.2.0
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
Flask==2.2.2
google-api-core==2.11.0
google-api-python-client==2.70.0
google-auth==2.15.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.57.0
gunicorn==20.1.0
httplib2==0.21.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
protobuf==4.21.12
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
requests==2.28.1
rsa==4.9
six==1.16.0
uritemplate==4.1.1
urllib3==1.26.13
Werkzeug==2.2.2